Make every decomp covering missing Meta register_meta, and test it
Added an API to let you query if there's already a registration for
a function, added a test to check that we register_meta for all
decompositions (except detach, that decomp is wrong lol), and then
update all the necessary sites to make the test pass.

Another strategy is to just unconditionally register_meta if there
isn't a meta registration.  Maybe that's better.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: ef0a69f8a17da0cc66eab9fc2bcdcfde20973f84
Pull Request resolved: https://github.com/pytorch/pytorch/pull/77386